backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.8.3
python-multipart==0.0.9
aiofiles==23.2.1